refactor: consolidate test files and improve organization
