--- __tests__/auth/protected-routes.test.tsx
@@ -1,43 +1,0 @@
-import { describe, test, expect, vi } from "vitest";
-
-// Mock the auth function for middleware testing
-const mockAuth = vi.fn();
-vi.mock("@/lib/auth", () => ({
-  auth: mockAuth,
-}));
-
-// Simple middleware logic test (since we can't easily test the actual middleware)
-describe("Protected Routes Logic", () => {
-  test("should redirect unauthenticated users from protected routes", () => {
-    const protectedPaths = ["/hiragana", "/katakana", "/dashboard"];
-    const isAuthenticated = false;
-
-    protectedPaths.forEach((path) => {
-      if (!isAuthenticated && path !== "/") {
-        // Should redirect to home
-        expect(path).not.toBe("/");
-      }
-    });
-  });
-
-  test("should allow authenticated users to access protected routes", () => {
-    const protectedPaths = ["/hiragana", "/katakana", "/dashboard"];
-    const isAuthenticated = true;
-
-    protectedPaths.forEach((path) => {
-      if (isAuthenticated) {
-        // Should allow access
-        expect(typeof path).toBe("string");
-      }
-    });
-  });
-
-  test("should allow unauthenticated access to home page", () => {
-    const homePath = "/";
-    const isAuthenticated = false;
-
-    // Home should always be accessible
-    expect(homePath).toBe("/");
-    expect(isAuthenticated).toBe(false);
-  });
-});
